+++ lib/src/null_exp.dart
@@ -1,0 +1,11 @@
+
+
+library bignum.null_exp;
+
+class NullExp {
+  NullExp();
+  convert(x) { return x; }
+  revert(x) { return x; }
+  mulTo(x,y,r) { x.multiplyTo(y,r); }
+  sqrTo(x,r) { x.squareTo(r); }
+}
